Update RobotPlayer.java
This is a new RobotPlayer file that we can put code into, separate from the example robot player. The current updates are that the miners will mark and remember the hq location at the start and the miners now will choose to mine soup first if they are able to.